workspace/requirement.txt
Pillow
PyMuPDF
pytesseract
python-docx
